Bring action response listeners into mixin for scalable reuse
